@@ -29,31 +29,31 @@
   "homepage": "https://github.com/foxdonut/meiosis-examples#readme",
   "dependencies": {
     "feather-route-matcher": "^3.1.0",
-    "marked": "^0.7.0",
-    "meiosis-routing": "^2.0.0",
-    "meiosis-setup": "^3.1.0",
+    "marked": "^0.8.2",
+    "meiosis-routing": "^2.0.1",
+    "meiosis-setup": "^4.0.0",
     "mergerino": "^0.4.0",
     "mithril": "^2.0.4",
-    "preact": "^10.0.4",
-    "query-string": "^6.8.3",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
+    "preact": "^10.3.4",
+    "query-string": "^6.11.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "seview": "^1.0.1",
     "tinyfunk": "^1.7.2",
     "validate.js": "^0.13.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "babel-loader": "^8.0.6",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-prettier": "^3.1.1",
+    "@babel/core": "^7.9.0",
+    "@babel/preset-env": "^7.9.0",
+    "babel-loader": "^8.1.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.1",
+    "eslint-plugin-prettier": "^3.1.2",
     "meiosis-tracer": "^3.0.0",
-    "prettier": "1.18.2",
-    "reload": "^3.0.3",
+    "prettier": "2.0.1",
+    "reload": "^3.0.4",
     "static-server": "^2.2.1",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10"
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11"
   }
 }
