@@ -1,4 +1,4 @@
-import React, { Component } from "react"
+import React from "react"
 
 import { conditions, Conditions } from "../conditions"
 import { dateTime, DateTime } from "../dateTime"
@@ -21,29 +21,21 @@ export const app = {
     )
 }
 
-export class App extends Component {
-  constructor(props) {
-    super(props)
-    this.state = props.states()
+export const App = ({ states, actions }) => {
+  const [init, setInit] = React.useState(false)
+  const [state, setState] = React.useState(states())
+
+  if (!init) {
+    setInit(true)
+    states.map(setState)
   }
 
-  componentDidMount() {
-    this.props.states.map(state => {
-      this.setState(state)
-    })
-  }
-
-  render() {
-    const state = this.state
-    const { actions } = this.props
-
-    return (
-      <div>
-        <DateTime state={state} id="dateTime" actions={actions} />
-        <Conditions state={state} id="conditions" actions={actions} />
-        <Temperature state={state} id="temperature:air" actions={actions} />
-        <Temperature state={state} id="temperature:water" actions={actions} />
-      </div>
-    )
-  }
+  return (
+    <div>
+      <DateTime state={state} id="dateTime" actions={actions} />
+      <Conditions state={state} id="conditions" actions={actions} />
+      <Temperature state={state} id="temperature:air" actions={actions} />
+      <Temperature state={state} id="temperature:water" actions={actions} />
+    </div>
+  )
 }
