@@ -24,7 +24,9 @@ module.exports = {
       {
         "printWidth": 100,
         "semi": false,
-        "singleQuote": false
+        "singleQuote": false,
+        "trailingComma": "none",
+        "arrowParens": "avoid"
       }
     ]
   },
