@@ -28,7 +28,9 @@ module.exports = {
       {
         "printWidth": 100,
         "semi": false,
-        "singleQuote": false
+        "singleQuote": false,
+        "trailingComma": "none",
+        "arrowParens": "avoid"
       }
     ],
     "react/display-name": [
